requirements.txt
# Core LangChain packages (compatible versions)
langchain>=0.3.7
langchain-openai>=0.2.9
langchain-community>=0.3.5
langchain-core>=0.3.15
langgraph>=0.2.45

# Search tool
tavily-python>=0.5.0
langchain-tavily>=0.2.0

# Checkpointer for LangGraph memory
langgraph-checkpoint-sqlite>=1.0.0

# Utilities
python-dotenv>=1.0.0

# Pydantic v2 (compatible with latest LangChain)
pydantic>=2.9.0,<3.0.0

# Testing
pytest>=8.3.0
